pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "sdi"
version = "0.14.0"
description = "Structural Divergence Indexer — measure and track structural drift in codebases"
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Geoff Godwin"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "tomli-w>=1.0",
    "ruamel.yaml>=0.18",
    "tomli>=2.0; python_version < '3.11'",
    "pathspec>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "ruff>=0.3",
    "mypy>=1.0",
]
docs = [
    "mkdocs>=1.6,<2",
    "mkdocs-material>=9.7,<10",
    "pymdown-extensions>=10.21,<11",
    "pygments>=2.20,<3",
]
all = [
    "tree-sitter>=0.24",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "tree-sitter-typescript",
    "tree-sitter-go",
    "tree-sitter-java",
    "tree-sitter-rust",
    "tree-sitter-bash",
    "leidenalg",
    "igraph",
]
web = []
systems = []

[project.scripts]
sdi = "sdi.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src"]
line-length = 120
extend-exclude = [
    "tests/fixtures",
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["tomli", "leidenalg", "igraph", "tree_sitter_bash"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "benchmark: performance benchmark tests (deselect with '-m \"not benchmark\"')",
]
